moved photos so they can be accessed as external data
The photo f3.jpg was the only photo not loading correctly as an
external image. I investigated and the only noticeable difference was
that f3.jpg contained a particularly large amount of metadata. On a
hunch I re-saved the file to remove the metadata.

This caused it to load properly.